mosaicme/web/requirements.txt
django-dotenv==1.3.0
redis==2.10.3
requests==2.6.0
boto==2.37.0
django==1.7.7
django-bootstrap3==5.4.0
django-angular==0.7.13
retrying==1.3.3
supervisor==3.1.3